Move index_options to options module for later refactoring
